refactor: settlementAccountForm 타입 모듈화, 일부 변수명 개선
